+++ components/tailored-import/back/src/Application/ExecuteDataMapping/UserIntentRegistry/UserIntentFactory/SimpleReferenceEntityUserIntentFactory.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Akeneo PIM Enterprise Edition.
+ *
+ * (c) 2022 Akeneo SAS (https://www.akeneo.com)
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Akeneo\Platform\TailoredImport\Application\ExecuteDataMapping\UserIntentRegistry\UserIntentFactory;
+
+use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\SetSimpleReferenceEntityValue;
+use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\ValueUserIntent;
+use Akeneo\Platform\TailoredImport\Application\ExecuteDataMapping\Exception\UnexpectedValueException;
+use Akeneo\Platform\TailoredImport\Application\ExecuteDataMapping\UserIntentRegistry\UserIntentFactoryInterface;
+use Akeneo\Platform\TailoredImport\Domain\Model\Target\AttributeTarget;
+use Akeneo\Platform\TailoredImport\Domain\Model\Target\TargetInterface;
+use Akeneo\Platform\TailoredImport\Domain\Model\Value\StringValue;
+use Akeneo\Platform\TailoredImport\Domain\Model\Value\ValueInterface;
+
+final class SimpleReferenceEntityUserIntentFactory implements UserIntentFactoryInterface
+{
+    /**
+     * @param AttributeTarget $target
+     */
+    public function create(TargetInterface $target, ValueInterface $value): ValueUserIntent
+    {
+        if (!$this->supports($target)) {
+            throw new \InvalidArgumentException('The target must be an AttributeTarget and be of type "akeneo_reference_entity"');
+        }
+
+        if (!$value instanceof StringValue) {
+            throw new UnexpectedValueException($value, StringValue::class, self::class);
+        }
+
+        return new SetSimpleReferenceEntityValue(
+            $target->getCode(),
+            $target->getChannel(),
+            $target->getLocale(),
+            $value->getValue(),
+        );
+    }
+
+    public function supports(TargetInterface $target): bool
+    {
+        return $target instanceof AttributeTarget && 'akeneo_reference_entity' === $target->getAttributeType();
+    }
+}